+++ cmd/forklift/env/cmd.go
@@ -57,7 +57,7 @@ var Cmd = &cli.Command{
 			Aliases:  []string{"cache-pallets"},
 			Category: "Use the environment",
 			Usage:    "Updates the cache with the pallets available in the local environment",
-			Action:   cachePalletAction,
+			Action:   cachePltAction,
 		},
 		{
 			Name:     "cache-img",
@@ -97,7 +97,7 @@ var Cmd = &cli.Command{
 			Aliases:  []string{"list-pallets"},
 			Category: "Query the environment",
 			Usage:    "Lists pallets available in the local environment",
-			Action:   lsPalletAction,
+			Action:   lsPltAction,
 		},
 		{
 			Name:      "show-plt",
@@ -105,7 +105,7 @@ var Cmd = &cli.Command{
 			Category:  "Query the environment",
 			Usage:     "Describes a pallet available in the local environment",
 			ArgsUsage: "pallet_path",
-			Action:    showPalletAction,
+			Action:    showPltAction,
 		},
 		{
 			Name:     "ls-pkg",