+++ internal/app/forklift/caching-models.go
@@ -20,39 +20,40 @@ type FSPkgLoader interface {
 	LoadFSPkgs(searchPattern string) ([]*pallets.FSPkg, error)
 }
 
-// Cache is a source of pallets and packages.
-type Cache interface {
+// PalletCache is a source of pallets and packages.
+type PalletCache interface {
 	FSPalletLoader
 	FSPkgLoader
 }
 
-// PathedCache is a Cache rooted at a single path.
-type PathedCache interface {
-	Cache
+// PathedPalletCache is a PalletCache rooted at a single path.
+type PathedPalletCache interface {
+	PalletCache
 	pallets.Pather
 }
 
-// FSCache is a local cache with copies of pallets (and thus of packages too), stored in a [fs.FS]
-// filesystem.
-type FSCache struct {
-	// FS is the filesystem which corresponds to the cache.
+// FSPalletCache is a local cache with copies of pallets (and thus of packages too), stored in a
+// [fs.FS] filesystem.
+type FSPalletCache struct {
+	// FS is the filesystem which corresponds to the cache of pallets.
 	FS pallets.PathedFS
 }
 
-// LayeredCache is a PathedCache where selected pallets can be overridden by another Cache, for
-// loading pallets and packages.
-// The path of the LayeredCache instance is just the path of the underlying cache.
-type LayeredCache struct {
+// LayeredPalletCache is a PathedPalletCache where selected pallets can be overridden by another
+// PalletCache, for loading pallets and packages.
+// The path of the LayeredPalletCache instance is just the path of the underlying cache.
+type LayeredPalletCache struct {
 	// Underlay is the underlying cache.
-	Underlay PathedCache
+	Underlay PathedPalletCache
 	// Overlay is the overlying cache which is used instead of the underlying cache for pallets and
 	// packages covered by the overlying cache.
-	Overlay OverlayCache
+	Overlay OverlayPalletCache
 }
 
-// OverlayCache is a cache which can report whether it includes any particular pallet or package.
-type OverlayCache interface {
-	Cache
+// OverlayPalletCache is a cache which can report whether it includes any particular pallet or
+// package.
+type OverlayPalletCache interface {
+	PalletCache
 	// IncludesFSPallet reports whether the cache expects to have the specified pallet.
 	// This result does not necessarily correspond to whether the cache actually has it.
 	IncludesFSPallet(palletPath string, version string) bool
@@ -61,10 +62,10 @@ type OverlayCache interface {
 	IncludesFSPkg(pkgPath string, version string) bool
 }
 
-// PalletOverrideCache is an [OverlayCache] implementation containing a set of pallets which can be
-// retrieved from the root of the cache. A pallet from the cache will be retrieved if it is stored
-// in the cache with a matching version, regardless of whether the pallet's own version actually
-// matches - in other words, pallets can be stored with fictional versions.
+// PalletOverrideCache is an [OverlayPalletCache] implementation containing a set of pallets which
+// can be retrieved from the root of the cache. A pallet from the cache will be retrieved if it is
+// stored in the cache with a matching version, regardless of whether the pallet's own version
+// actually matches - in other words, pallets can be stored with fictional versions.
 type PalletOverrideCache struct {
 	// pallets is a map associating pallet paths to loaded pallets.
 	// For each key-value pair, the key must be the path of the pallet which is the value of that