+++ internal/app/forklift/cli/env-depls.go
@@ -29,7 +29,7 @@ func PrintEnvDepls(indent int, env *forklift.FSEnv, loader forklift.FSPkgLoader)
 }
 
 func PrintDeplInfo(
-	indent int, env *forklift.FSEnv, cache forklift.PathedCache, deplName string,
+	indent int, env *forklift.FSEnv, cache forklift.PathedPalletCache, deplName string,
 ) error {
 	depl, err := env.LoadDepl(deplName)
 	if err != nil {
@@ -60,7 +60,7 @@ func PrintDeplInfo(
 	return nil
 }
 
-func printDepl(indent int, cache forklift.PathedCache, depl *forklift.ResolvedDepl) {
+func printDepl(indent int, cache forklift.PathedPalletCache, depl *forklift.ResolvedDepl) {
 	IndentedPrintf(indent, "Package deployment: %s\n", depl.Name)
 	indent++
 
@@ -86,7 +86,7 @@ func printDepl(indent int, cache forklift.PathedCache, depl *forklift.ResolvedDe
 	printFeatures(indent+1, disabledFeatures)
 }
 
-func printDeplPkg(indent int, cache forklift.PathedCache, depl *forklift.ResolvedDepl) {
+func printDeplPkg(indent int, cache forklift.PathedPalletCache, depl *forklift.ResolvedDepl) {
 	IndentedPrintf(indent, "Deploys package: %s\n", depl.Def.Package)
 	indent++